add pooling layer
